add AboutUnits component, add more margin to jumping arrow
